+++ R/tableby.control.R
@@ -38,7 +38,7 @@
 #' @param digits.count Number of decimal places for count values.
 #' @param digits.pct Number of decimal places for percents.
 #' @param digits.p Number of decimal places for p-values.
-#' @param format.p Logical, denoting whether to format p-values. See "Details", below.
+#' @param format.p Logical, denoting whether to format p-values, or character, a \code{\link[glue]{glue}} specification for how to format. See "Details", below.
 #' @param digits.n Number of decimal places for N's in the header. Set it to NA to suppress the N's.
 #' @param conf.level Numeric, denoting what confidence level to use for confidence intervals.
 #'   (See, e.g., \code{\link{binomCI}})
@@ -53,7 +53,8 @@
 #' If \code{format.p} is \code{FALSE}, \code{digits.p} denotes the number of significant digits shown. The
 #'   p-values will be in exponential notation if necessary. If \code{format.p} is \code{TRUE},
 #'   \code{digits.p} will determine the number of digits after the decimal point to show. If the p-value
-#'   is less than the resulting number of places, it will be formatted to show so.
+#'   is less than the resulting number of places, it will be formatted to show so. If \code{format.p} is a character string,
+#'   it will be treated as a \code{\link[glue]{glue}} specification: the p-value is exposed as "p", and "digits.p" as "digits.p".
 #'
 #' Options for statistics are described more thoroughly in the vignette and are listed in \link{tableby.stats}
 #'